+++ Libplanet/Blockchain/BlockChain.cs
@@ -19,6 +19,15 @@ using Libplanet.Tx;
 [assembly: InternalsVisibleTo("Libplanet.Tests")]
 namespace Libplanet.Blockchain
 {
+    /// <summary>
+    /// A class have <see cref="Block{T}"/>s, <see cref="Transaction{T}"/>s, and the chain
+    /// information.
+    /// </summary>
+    /// <typeparam name="T">An <see cref="IAction"/> type.  It should match
+    /// to <see cref="Block{T}"/>'s type parameter.</typeparam>
+    /// <seealso cref="IAction"/>
+    /// <seealso cref="Block{T}"/>
+    /// <seealso cref="Transaction{T}"/>
     public class BlockChain<T> : IReadOnlyList<Block<T>>
         where T : IAction, new()
     {
@@ -30,6 +39,29 @@ namespace Libplanet.Blockchain
         internal readonly ReaderWriterLockSlim _rwlock;
         private readonly object _txLock;
 
+        /// <summary>
+        /// All <see cref="Block{T}"/>s in the <see cref="BlockChain{T}"/>
+        /// storage, including orphan <see cref="Block{T}"/>s.
+        /// Keys are <see cref="Block{T}.Hash"/>es and values are
+        /// their corresponding <see cref="Block{T}"/>s.
+        /// </summary>
+        private IDictionary<HashDigest<SHA256>, Block<T>> _blocks;
+
+        /// <summary>
+        /// All <see cref="Transaction{T}"/>s in the <see cref="BlockChain{T}"/>
+        /// storage, including orphan <see cref="Transaction{T}"/>s.
+        /// Keys are <see cref="Transaction{T}.Id"/>s and values are
+        /// their corresponding <see cref="Transaction{T}"/>s.
+        /// </summary>
+        private IDictionary<TxId, Transaction<T>> _transactions;
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="BlockChain{T}"/> class.
+        /// </summary>
+        /// <param name="policy"><see cref="IBlockPolicy{T}"/> to use in the
+        /// <see cref="BlockChain{T}"/>.</param>
+        /// <param name="store"><see cref="IStore"/> to store <see cref="Block{T}"/>s,
+        /// <see cref="Transaction{T}"/>s, and <see cref="BlockChain{T}"/> information.</param>
         public BlockChain(IBlockPolicy<T> policy, IStore store)
             : this(
                 policy,
@@ -44,9 +76,9 @@ namespace Libplanet.Blockchain
             Id = id;
             Policy = policy;
             Store = store;
-            Blocks = new BlockSet<T>(store);
-            Transactions = new TransactionSet<T>(store);
 
+            _blocks = new BlockSet<T>(store);
+            _transactions = new TransactionSet<T>(store);
             _rwlock = new ReaderWriterLockSlim(LockRecursionPolicy.SupportsRecursion);
             _txLock = new object();
 
@@ -90,28 +122,6 @@ namespace Libplanet.Blockchain
         public Guid Id { get; private set; }
 
         /// <summary>
-        /// All <see cref="Block{T}"/>s in the <see cref="BlockChain{T}"/>
-        /// storage, including orphan <see cref="Block{T}"/>s.
-        /// Keys are <see cref="Block{T}.Hash"/>es and values are
-        /// their corresponding <see cref="Block{T}"/>s.
-        /// </summary>
-        public IDictionary<HashDigest<SHA256>, Block<T>> Blocks
-        {
-            get; private set;
-        }
-
-        /// <summary>
-        /// All <see cref="Transaction{T}"/>s in the <see cref="BlockChain{T}"/>
-        /// storage, including orphan <see cref="Transaction{T}"/>s.
-        /// Keys are <see cref="Transaction{T}.Id"/>s and values are
-        /// their corresponding <see cref="Transaction{T}"/>s.
-        /// </summary>
-        public IDictionary<TxId, Transaction<T>> Transactions
-        {
-            get; private set;
-        }
-
-        /// <summary>
         /// All <see cref="Block{T}.Hash"/>es in the current index.  The genesis block's hash goes
         /// first, and the tip goes last.
         /// </summary>
@@ -123,29 +133,129 @@ namespace Libplanet.Blockchain
 
         internal IStore Store { get; }
 
-        /// <inheritdoc/>
+        /// <summary>
+        /// Gets the block corresponding to the <paramref name="index"/>.
+        /// </summary>
+        /// <param name="index">A number of index of <see cref="Block{T}"/>.</param>
+        /// <exception cref="ArgumentOutOfRangeException">Thrown when the given index of
+        /// <see cref="Block{T}"/> does not exist.</exception>
         public Block<T> this[int index] => this[(long)index];
 
+        /// <summary>
+        /// Gets the block corresponding to the <paramref name="index"/>.
+        /// </summary>
+        /// <param name="index">A number of index of <see cref="Block{T}"/>.</param>
+        /// <exception cref="ArgumentOutOfRangeException">Thrown when the given index of
+        /// <see cref="Block{T}"/> does not exist.</exception>
         public Block<T> this[long index]
         {
             get
             {
+                _rwlock.EnterReadLock();
                 try
                 {
-                    _rwlock.EnterReadLock();
-
                     HashDigest<SHA256>? blockHash = Store.IndexBlockHash(Id, index);
                     if (blockHash == null)
                     {
                         throw new ArgumentOutOfRangeException();
                     }
 
-                    return Blocks[blockHash.Value];
+                    return _blocks[blockHash.Value];
                 }
                 finally
                 {
                     _rwlock.ExitReadLock();
                 }
+            }
+        }
+
+        /// <summary>
+        /// Gets the block corresponding to the <paramref name="hash"/>.
+        /// </summary>
+        /// <param name="hash">A <see cref="Block{T}.Hash"/> of the <see cref="Block{T}"/> to get.
+        /// </param>
+        /// <exception cref="KeyNotFoundException">Thrown when there is no <see cref="Block{T}"/>
+        /// with a given <paramref name="hash"/>.</exception>
+        public Block<T> this[HashDigest<SHA256> hash]
+        {
+            get
+            {
+                _rwlock.EnterReadLock();
+                try
+                {
+                    return _blocks[hash];
+                }
+                finally
+                {
+                    _rwlock.ExitReadLock();
+                }
+            }
+        }
+
+        /// <summary>
+        /// Determines whether the <see cref="BlockChain{T}"/> contains <see cref="Block{T}"/>
+        /// the specified <paramref name="hash"/>.
+        /// </summary>
+        /// <param name="hash">The <see cref="HashDigest{T}"/> of the <see cref="Block{T}"/> to
+        /// check if it is in the <see cref="BlockChain{T}"/>.</param>
+        /// <returns>
+        /// <c>true</c> if the <see cref="BlockChain{T}"/> contains <see cref="Block{T}"/> with
+        /// the specified <paramref name="hash"/>; otherwise, <c>false</c>.
+        /// </returns>
+        public bool Contains(HashDigest<SHA256> hash)
+        {
+            _rwlock.EnterReadLock();
+            try
+            {
+                return _blocks.ContainsKey(hash);
+            }
+            finally
+            {
+                _rwlock.ExitReadLock();
+            }
+        }
+
+        /// <summary>
+        /// Determines whether the <see cref="BlockChain{T}"/> contains <see cref="Transaction{T}"/>
+        /// the specified <paramref name="txId"/>.
+        /// </summary>
+        /// <param name="txId">The <see cref="HashDigest{T}"/> of the <see cref="Transaction{T}"/>
+        /// to check if it is in the <see cref="BlockChain{T}"/>.</param>
+        /// <returns>
+        /// <c>true</c> if the <see cref="BlockChain{T}"/> contains <see cref="Transaction{T}"/>
+        /// with the specified <paramref name="txId"/>; otherwise, <c>false</c>.
+        /// </returns>
+        public bool Contains(TxId txId)
+        {
+            _rwlock.EnterReadLock();
+            try
+            {
+                return _transactions.ContainsKey(txId);
+            }
+            finally
+            {
+                _rwlock.ExitReadLock();
+            }
+        }
+
+        /// <summary>
+        /// Gets the transaction corresponding to the <paramref name="txId"/>.
+        /// </summary>
+        /// <param name="txId">A <see cref="TxId"/> of the <see cref="Transaction{T}"/> to get.
+        /// </param>
+        /// <returns><see cref="Transaction{T}"/> with <paramref name="txId"/>.</returns>
+        /// <exception cref="KeyNotFoundException">Thrown when there is no
+        /// <see cref="Transaction{T}"/> with a given <paramref name="txId"/>.</exception>
+        public Transaction<T> GetTransaction(TxId txId)
+        {
+            _rwlock.EnterReadLock();
+            try
+            {
+                return _transactions[txId];
+            }
+            finally
+            {
+                _rwlock.ExitReadLock();
             }
         }
 
@@ -173,7 +283,7 @@ namespace Libplanet.Blockchain
 
         public IEnumerator<Block<T>> GetEnumerator()
         {
-            return BlockHashes.Select(hash => Blocks[hash]).GetEnumerator();
+            return BlockHashes.Select(hash => _blocks[hash]).GetEnumerator();
         }
 
         IEnumerator IEnumerable.GetEnumerator()
@@ -235,7 +345,7 @@ namespace Libplanet.Blockchain
                 return states;
             }
 
-            Block<T> block = Blocks[offset.Value];
+            Block<T> block = this[offset.Value];
             Tuple<HashDigest<SHA256>, long> stateReference;
 
             _rwlock.EnterReadLock();
@@ -370,7 +480,7 @@ namespace Libplanet.Blockchain
             {
                 foreach (Transaction<T> tx in transactions)
                 {
-                    Transactions[tx.Id] = tx;
+                    _transactions[tx.Id] = tx;
                 }
 
                 Store.StageTransactionIds(transactions.Select(tx => tx.Id).ToImmutableHashSet());
@@ -602,7 +712,7 @@ namespace Libplanet.Blockchain
                 try
                 {
                     Block<T> prevTip = Tip;
-                    Blocks[block.Hash] = block;
+                    _blocks[block.Hash] = block;
                     foreach (KeyValuePair<Address, long> pair in nonceDeltas)
                     {
                         Store.IncreaseTxNonce(Id, pair.Key, pair.Value);
@@ -756,8 +866,8 @@ namespace Libplanet.Blockchain
 
                 foreach (HashDigest<SHA256> hash in locator)
                 {
-                    if (Blocks.ContainsKey(hash)
-                        && Blocks[hash] is Block<T> block
+                    if (_blocks.ContainsKey(hash)
+                        && _blocks[hash] is Block<T> block
                         && hash.Equals(Store.IndexBlockHash(Id, block.Index)))
                     {
                         return hash;
@@ -789,7 +899,7 @@ namespace Libplanet.Blockchain
 
                 HashDigest<SHA256>? branchPoint = FindBranchPoint(locator);
                 var branchPointIndex = branchPoint is HashDigest<SHA256> h
-                    ? (int)Blocks[h].Index
+                    ? (int)_blocks[h].Index
                     : 0;
 
                 // FIXME: Currently, increasing count by one to satisfy
@@ -837,7 +947,7 @@ namespace Libplanet.Blockchain
                 _rwlock.EnterReadLock();
 
                 Store.ForkBlockIndexes(Id, forkedId, point);
-                Block<T> pointBlock = Blocks[point];
+                Block<T> pointBlock = _blocks[point];
 
                 var signersToStrip = new Dictionary<Address, int>();
 
@@ -845,7 +955,7 @@ namespace Libplanet.Blockchain
                     Block<T> block = Tip;
                     block.PreviousHash is HashDigest<SHA256> hash
                     && !block.Hash.Equals(point);
-                    block = Blocks[hash])
+                    block = _blocks[hash])
                 {
                     IEnumerable<(Address, int)> signers = block
                         .Transactions
@@ -908,7 +1018,7 @@ namespace Libplanet.Blockchain
                 while (current is HashDigest<SHA256> hash)
                 {
                     hashes.Add(hash);
-                    Block<T> currentBlock = Blocks[hash];
+                    Block<T> currentBlock = _blocks[hash];
 
                     if (currentBlock.Index == 0)
                     {
@@ -947,7 +1057,7 @@ namespace Libplanet.Blockchain
                     Block<T> t = this[shorterHeight], o = other[shorterHeight];
                     t.PreviousHash is HashDigest<SHA256> tp &&
                         o.PreviousHash is HashDigest<SHA256> op;
-                    t = Blocks[tp], o = other.Blocks[op]
+                    t = this[tp], o = other[op]
                 )
                 {
                     if (t.Equals(o))
@@ -965,7 +1075,7 @@ namespace Libplanet.Blockchain
                     Block<T> b = Tip;
                     !(b is null) && b.Index > (topmostCommon?.Index ?? -1) &&
                         b.PreviousHash is HashDigest<SHA256> ph;
-                    b = Blocks[ph]
+                    b = this[ph]
                 )
                 {
                     List<ActionEvaluation> evaluations = b.EvaluateActionsPerTx(a =>
@@ -1003,9 +1113,9 @@ namespace Libplanet.Blockchain
                 Guid obsoleteId = Id;
                 Id = other.Id;
                 Store.SetCanonicalChainId(Id);
-                Blocks = new BlockSet<T>(Store);
+                _blocks = new BlockSet<T>(Store);
                 TipChanged?.Invoke(this, tipChangedEventArgs);
-                Transactions = new TransactionSet<T>(Store);
+                _transactions = new TransactionSet<T>(Store);
                 Store.DeleteChainId(obsoleteId);
             }
             finally
@@ -1101,7 +1211,7 @@ namespace Libplanet.Blockchain
             {
                 foreach (HashDigest<SHA256> hash in IterateBlockHashes(offset, limit))
                 {
-                    yield return Blocks[hash];
+                    yield return _blocks[hash];
                 }
             }
             finally