+++ CHANGES.md
@@ -24,10 +24,16 @@ To be released.
         became replaced by `LoadPlainValue(IValue)`.
      -  `AccountStateGetter` became to return `IValue`, not `object`.
      -  Added `BencodexExtension` static class.
+ -  Removed `BlockChain<T>.Blocks`.  [[#409], [#583]]
+ -  Removed `BlockChain<T>.Transactions`.  [[#409], [#583]]
 
 ### Added interfaces
 
  -  Added `BlockChain<T>.LongCount()` method.  [[#575]]
+ -  Added `BlockChain<T>[HashDigest<T>]` indexer.  [[#409], [#583]]
+ -  Added `BlockChain<T>.Contains(HashDigest<T>)` method.  [[#409], [#583]]
+ -  Added `BlockChain<T>.GetTransaction(TxId)` method.  [[#409], [#583]]
+ -  Added `BlockChain<T>.Contains(TxId)` method.  [[#409], [#583]]
 
 ### Behavioral changes
 
@@ -54,6 +60,7 @@ To be released.
 
 [#209]: https://github.com/planetarium/libplanet/issues/209
 [#405]: https://github.com/planetarium/libplanet/issues/405
+[#409]: https://github.com/planetarium/libplanet/issues/409
 [#447]: https://github.com/planetarium/libplanet/issues/447
 [#462]: https://github.com/planetarium/libplanet/issues/462
 [#469]: https://github.com/planetarium/libplanet/pull/469
@@ -73,6 +80,7 @@ To be released.
 [#566]: https://github.com/planetarium/libplanet/pull/566
 [#575]: https://github.com/planetarium/libplanet/pull/575
 [#576]: https://github.com/planetarium/libplanet/pull/576
+[#583]: https://github.com/planetarium/libplanet/pull/583
 
 
 Version 0.6.0